Add support for Android saves
Android save support now working.

Android saves use version 0x13 for the save file (vs 0x10 for the PC
version of EW). The Android saves differ from PC in several key ways:

- Android saves have no header checksum, just the compressed data
checksum. It occurs in the same place as for PC, and is the same CRC32B
algorithm.

- The compressed data section of Android saves are compressed using zlib
instead of lzo.

- There is also a small bit of data after the checksum in the header but
before the compressed data. This may be Android profile related?

+++ xcom2json.cpp
@@ -528,7 +528,7 @@ void buildJson(const saved_game& save, json_writer& w)
 
     w.write_key("header");
     w.begin_object();
-    w.write_int("version", hdr.version);
+    w.write_int("version", static_cast<uint32_t>(hdr.version));
     w.write_int("uncompressed_size", hdr.uncompressed_size);
     w.write_int("game_number", hdr.game_number);
     w.write_int("save_number", hdr.save_number);
